GH-2291 Merge tests into one file
